IGNITE-22266 Move IgniteTables to org.apache.ignite.table package
https://issues.apache.org/jira/browse/IGNITE-22266
