+++ modules/client/src/main/java/org/apache/ignite/internal/client/table/ClientTables.java
@@ -27,7 +27,7 @@ import org.apache.ignite.internal.client.ReliableChannel;
 import org.apache.ignite.internal.client.proto.ClientOp;
 import org.apache.ignite.internal.marshaller.MarshallersProvider;
 import org.apache.ignite.table.Table;
-import org.apache.ignite.table.manager.IgniteTables;
+import org.apache.ignite.table.IgniteTables;
 
 /**
  * Client tables API implementation.
