+++ modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/api/ItRecordViewKeyPositionsTest.java
@@ -30,7 +30,7 @@ import org.apache.ignite.internal.tostring.IgniteToStringInclude;
 import org.apache.ignite.internal.tostring.S;
 import org.apache.ignite.table.RecordView;
 import org.apache.ignite.table.Tuple;
-import org.apache.ignite.table.manager.IgniteTables;
+import org.apache.ignite.table.IgniteTables;
 import org.apache.ignite.table.mapper.Mapper;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Named;
